pyproject.toml
[project]
name = "urban-flow"
version = "0.1.0"
description = "Comprehensive traffic forecasting system for urban junctions"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=2.3.3",
    "pmdarima>=2.1.1",
    "properscoring>=0.1",
    "pytorch-forecasting>=1.6.1",
    "pytorch-lightning>=2.6.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.6",
    "torch>=2.10.0",
    "xgboost>=3.2.0",
]
